Fix settle: transfer payment as coins split out from game balance
Sui does not allow contract to take ownership of users' coins unless
the transaction is signed by the users.  Settlement transactions are
signed only by the transactor so no user's coin can be used.

Instead, the payment coin is created by splitting out the needed
balance from the game deposits and make a coin out of the balance.
The coin then is transferred to the settle player accordingly.